Backed out changeset 7793fa575497 (bug 1580087) for build bustages at nsIconChannel.cpp. CLOSED TREE
UltraBlame original commit: 0879b99736d84ff99ebc013fa30dabe1da067b36

+++ image/ClippedImage.cpp
@@ -512,8 +512,7 @@ nsIntSize ClippedImage::OptimalImageSizeForDest(const gfxSize& aDest,
 
     
     
-    gfxSize desiredSize(double(imgWidth) * scale.width,
-                        double(imgHeight) * scale.height);
+    gfxSize desiredSize(imgWidth * scale.width, imgHeight * scale.height);
     nsIntSize innerDesiredSize = InnerImage()->OptimalImageSizeForDest(
         desiredSize, aWhichFrame, aSamplingFilter, aFlags);
